Long Overdue Commit
About a year later, making this final commit. Generally, the app was
working fine. It was designed to run on a Samsung Galaxy either Tab or
Note 8.0, I believe. It worked fine but had some issues crashing when
pulling up a couple of the photos. I never looked into finishing
debugging that. Recently installed the app on an LG G4; the app doesn't
look right on the smaller screen, and still has the same crashing
issues. Will be starting a new version of the app sometime soon.